+++ src/timecard/interface/timedisplay.py
@@ -7,7 +7,7 @@ and state.
 
 from datetime import datetime
 
-from PySide2.QtCore import QDateTime, QElapsedTimer, QTimer
+from PySide2.QtCore import QDateTime, QDate, QTime, QElapsedTimer, QTimer
 from PySide2.QtWidgets import QLCDNumber
 
 
@@ -17,10 +17,14 @@ class TimeDisplay:
     timer = None
     time = None
     timestamp = None
+    # Stored time in ms from prior pause(s)
     freeze = 0
     paused = False
 
-    callback = None
+    from_backup = False
+
+    callback_tick = []
+    callback_stop = []
 
     @classmethod
     def build(cls):
@@ -32,21 +36,45 @@ class TimeDisplay:
         return cls.lcd
 
     @classmethod
+    def restore_from_backup(cls, timestamp: datetime, elapsed):
+        """Restore the timer state from backup."""
+        # Just ignore less than a second of time.
+        if elapsed < (1000):
+            return
+
+        # If loading from backup, indicate that fact so we don't clobber timestamp.
+        cls.from_backup = True
+        # Restore the timestamp.
+        cls.timestamp = QDateTime(
+            QDate(timestamp.year, timestamp.month, timestamp.day),
+            QTime(timestamp.hour, timestamp.minute, timestamp.second)
+        )
+        # Restore the duration
+        cls.freeze = elapsed
+        # Force the user to either save or reset. Resuming makes no sense.
+        cls.paused = False
+        # Show the time.
+        cls.show_time(*cls.get_time(), silent=True)
+
+    @classmethod
     def show_default(cls):
         """Reset the time display to 00:00:00"""
         cls.show_time()
 
     @classmethod
-    def show_time(cls, hours=0, minutes=0, seconds=0):
+    def show_time(cls, hours=0, minutes=0, seconds=0, silent=False):
         """Show the indicated time on the GUI.
 
         hours -- the number of hours
         minutes -- the number of minutes
         seconds -- the number of seconds
+
+        silent -- don't fire callbacks
         """
         cls.lcd.display(f"{hours:02}:{minutes:02}:{seconds:02}")
-        if cls.callback:
-            cls.callback(hours, minutes, seconds)
+        if not silent:
+            for callback in cls.callback_tick:
+                callback(hours, minutes, seconds)
 
     @classmethod
     def update_time(cls):
@@ -62,10 +90,18 @@ class TimeDisplay:
         if cls.time is None:
             cls.time = QElapsedTimer()
             cls.time.start()
-            cls.timestamp = QDateTime.currentDateTime()
+
+            # Do not clobber timestamp when restoring from backup
+            if not cls.from_backup:
+                cls.timestamp = QDateTime.currentDateTime()
+            else:
+                # After restoring from backup, forget about it.
+                # We need to be able to create new timestamps afterward.
+                cls.from_backup = False
         else:
             cls.time.restart()
-        cls.timer.start()
+
+        cls.timer.start(1000)  # tick once every second (performance!)
         cls.paused = False
 
     @classmethod
@@ -90,6 +126,8 @@ class TimeDisplay:
         cls.time = None
         if erase:
             cls.freeze = 0
+        for callback in cls.callback_stop:
+            callback(erase)
 
     @classmethod
     def get_time_ms(cls):
@@ -97,6 +135,10 @@ class TimeDisplay:
         if cls.time:
             return cls.time.elapsed() + cls.freeze
         return cls.freeze
+
+    @classmethod
+    def get_time_ms_string(cls):
+        return str(cls.get_time_ms())
 
     @classmethod
     def get_time(cls):
@@ -128,6 +170,28 @@ class TimeDisplay:
         return timestamp
 
     @classmethod
-    def subscribe(cls, callback=None):
+    def get_timestamp_string(cls):
+        """Returns the timestamp as a storable string."""
+        return (f"{cls.timestamp.date().year()}:"
+                f"{cls.timestamp.date().month()}:"
+                f"{cls.timestamp.date().day()}:"
+                f"{cls.timestamp.time().hour()}:"
+                f"{cls.timestamp.time().minute()}:"
+                f"{cls.timestamp.time().second()}"
+                )
+
+    @classmethod
+    def subscribe_tick(cls, callback=None):
+        """Subscribe to every clock tick
+        Callback callable must accept the arguments: (hours, minutes, seconds)
+        """
         if callback:
-            cls.callback = callback
+            cls.callback_tick.append(callback)
+
+    @classmethod
+    def subscribe_stop(cls, callback=None):
+        """Subscribe to the clock being stopped (not paused)
+        Callback callable must accept the arguments: (erased: bool)
+        """
+        if callback:
+            cls.callback_stop.append(callback)