+++ CHANGELOG.md
@@ -2,12 +2,24 @@
 
 ## [Unreleased]
 
+### Added
+
+- Lost Timer Recovery: Saves a backup of timer state every minute in case of crash or accidental shutdown.
+
+  - Does not load a running instance's backups, so you can still have multiple running instances.
+
 ### Changed
 
+- Significant performance improvement! Now uses less then 5% CPU after startup, often less than 1%.
+
 - Timecard settings file now supports line comments. Line must start with `#` to be a comment.
+
 - Deprecate `.timecardrc` in favor of user folders (e.g. `.config/timecard/settings.conf`).
+
   - If new default settings file doesn't exist, will check for `.timecardrc` before creating new default.
+
   - If loading from `.timecardrc`, will migrate settings to new location and add migration comments to old location.
+
 - Default location for time logs is now the user data directory (e.g. `.local/share/timecard/`).
 
 ## [2.0.7] - 2021-03-24